What is Dingding?
I have no clue what a Dingding message is or what a Dingding operator is. I did a Google search and it seems as if Alibaba has some sort of messaging service. Is this what it is? This should be explained briefly on the page, so I can decide if it is something I need to understand or can happily ignore.